fix(web-ext): タブトラッキングを useNavigate ベースに変更
window.location.hash の直接操作を React Router の useNavigate に
置き換え、ナビゲーション履歴を正しく管理する。
components/activeTab モジュールとして切り出し、App コンポーネント
から呼び出すよう変更。
