+++ apps/web-ext/src/components/activeTab/index.ts
@@ -1,0 +1,1 @@
+export { useTabTracking } from "./use-tab-tracking";
